command.js renamed to app.js
