+++ temp.html
@@ -1,0 +1,20 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <title>Dolos summary</title>
+</head>
+
+<body>
+  
+    <table>
+      <tr>
+        <th>File 1</th>
+        <th>File 2</th>
+        <th>Lines matched</th>
+      </tr>
+      <tr><td>samples/js/sample.js</td><td>samples/js/another_copied_function.js</td><td  align="right">20</td></tr><tr><td>samples/js/copied_function.js</td><td>samples/js/copied_function.js</td><td  align="right">12</td></tr><tr><td>samples/js/copied_function.js</td><td>samples/js/another_copied_function.js</td><td  align="right">2</td></tr>
+    </table>
+    
+</body>
+</html>
